+++ lib/sidekiq/instrumental/middleware/server.rb
@@ -18,7 +18,7 @@ module Sidekiq
           gauge(base_key + 'time', elapsed)
           gauge(base_key + 'enqueued', stats.queues[queue].to_i)
           gauge(base_key + 'latency', Sidekiq::Queue.new(queue.to_s).latency)
-          base_key += msg.display_class.underscore.gsub('/', '_') + '.'
+          base_key += build_class_key(msg.display_class) + '.'
 
           increment(base_key + 'processed')
           gauge(base_key + 'time', elapsed)