+++ spec/sidekiq/instrumental/middleware/server_spec.rb
@@ -91,6 +91,41 @@ RSpec.describe Sidekiq::Instrumental::Middleware::Server do
     subject
   end
 
+  describe 'metric naming' do
+    context 'when the class name contains unsupported characters' do
+      let(:msg) { { 'class' => 'MyMailer#message/fun' } }
+
+      it 'changes them to _' do
+        expect(middleware).to receive(:increment)
+                                .with("sidekiq.#{queue}.my_mailer_message_fun.processed")
+
+        subject
+      end
+    end
+
+    context 'when the class name causes more then one . in a row' do
+      let(:msg) { { 'class' => 'MyMailer..message' } }
+
+      it 'combines them to one' do
+        expect(middleware).to receive(:increment)
+                                .with("sidekiq.#{queue}.my_mailer.message.processed")
+
+        subject
+      end
+    end
+
+    context 'trims a trailing .' do
+      let(:msg) { { 'class' => 'MyMailer#message.' } }
+
+      it 'combines them to one' do
+        expect(middleware).to receive(:increment)
+                                .with("sidekiq.#{queue}.my_mailer_message.processed")
+
+        subject
+      end
+    end
+  end
+
   it 'gauge elapsed time metric for the class' do
     expect(middleware).to receive(:gauge)
                             .with(