+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 ### Changed
 ### Fixed
+- class name calculation needs to filter out unsupported characters by instrumental
 
 ## [0.3.0]
 ### Fixed